fix(desktop): extract control-plane param types from the frozen backend shell
Keep desktop_backend.rs under the v0.36.26 baseline so the split gate stays green. The serde types the pipe schemas use are unchanged.
