+++ scripts/test-v03626-desktop-split-gate.ps1
@@ -198,6 +198,7 @@ $moduleBudgets = @(
     [ordered]@{ name = 'desktop frontend entrypoint'; path = 'winsmux-app/src/main.ts'; baseline = 19172; allowed_growth = 200; require_not_above_baseline = $true },
     [ordered]@{ name = 'rust operator cli'; path = 'core/src/operator_cli.rs'; baseline = 11658; allowed_growth = 199; require_not_above_baseline = $true },
     [ordered]@{ name = 'desktop backend shell'; path = 'winsmux-app/src-tauri/src/desktop_backend.rs'; baseline = 6152; allowed_growth = 56; require_not_above_baseline = $true },
+    [ordered]@{ name = 'desktop control-plane params'; path = 'winsmux-app/src-tauri/src/desktop_control_plane_params.rs'; baseline = 158; allowed_growth = 40; require_not_above_baseline = $true },
     [ordered]@{ name = 'desktop team profile bridge'; path = 'winsmux-app/src-tauri/src/desktop_team_profile.rs'; baseline = 210; allowed_growth = 40; require_not_above_baseline = $true }
 )
