+++ winsmux-app/src-tauri/src/lib.rs
@@ -1,5 +1,6 @@
 mod control_pipe;
 mod desktop_backend;
+mod desktop_control_plane_params;
 mod desktop_companion_cli;
 mod desktop_events;
 mod desktop_session_restore;
